95-96 | Pause Menu contd | (bug in BP see note)
Boolean function on Main CanMove to check whether Main should be able to move or not.
On video 96 in the Widget for Pause Menu BP we are trying to program the buttons in BP.  There seems to be a bug and the C++ functions cannot be called after casting to MainPlayerController.  As a result the pause menu is incomplete